renamed stereo_synced file and added doc
